src/ltu/requirements.txt
accelerate==0.23.0
aiohttp==3.8.5
altair==5.1.2
datasets==2.14.5
docker-pycreds==0.4.0
fastapi==0.103.2
ffmpeg-python==0.2.0
fire==0.5.0
fonttools==4.43.1
fsspec==2023.6.0
future==0.18.3
GitPython==3.1.37
gradio==3.47.1
h11==0.14.0
httpx==0.25.0
huggingface-hub==0.17.3
imageio==2.31.5
Jinja2==3.1.2
jsonschema==4.19.1
matplotlib==3.8.0
numba==0.58.0
numpy==1.23.5
openai_whisper
orjson==3.9.7
pandas==2.1.1
Pillow
protobuf==4.24.4
psutil==5.9.5
pyarrow==13.0.0
pydantic==2.4.2
PyYAML==6.0.1
regex==2023.10.3
requests
scikit-image==0.22.0
scipy==1.11.3
sentencepiece==0.1.99
sentry-sdk==1.31.0
six==1.16.0
sniffio==1.3.0
starlette==0.27.0
tifffile==2023.9.26
timm==0.4.5
tokenizers==0.13.3
torch==1.13.1
torchaudio==0.13.1
torchvision==0.14.1
transformers
triton==2.0.0
uvicorn==0.23.2
wandb==0.15.12
websockets==11.0.3
whisper-at==0.5